# Remove explicit casts.