@eloj 2024-11-03
Ran the matrix of clang++-19 and GCC 14.2 before and after this change using the timing branch (#11), no special prep: 

```
before-gcc.txt:Frame 8 time: 1529 ms, avg: 1531.00 ms
after-gcc.txt:Frame 8 time: 1531 ms, avg: 1529.00 ms
before-clang.txt:Frame 8 time: 1352 ms, avg: 1348.75 ms
after-clang.txt:Frame 8 time: 1339 ms, avg: 1339.38 ms
```