@jbikker 2024-11-03
Removing the casts is fine and makes the code more compact and clear. Originally I had that TriVertex thing in there but that was silly, bvhvec4 is just fine and has automatic conversion to and from bvhvec3.